refactor: 정적팩터리메서드 추가
1. SymbolToPieceConvert 클래스 삭제
2. SymbolToPieceConvert클래스에서 하는 기능을 정적팩터리 메서드가 수행
