+++ src/main/java/chess/domain/piece/AbstractPiece.java
@@ -8,13 +8,50 @@ import java.util.Objects;
 public abstract class AbstractPiece implements Piece {
 
     protected static final String ERROR_CAN_NOT_MOVE = "기물이 이동할 수 없는 위치입니다.";
+    private static final String PAWN = "p";
+    private static final String ROOK = "r";
+    private static final String KNIGHT = "n";
+    private static final String QUEEN = "q";
+    private static final String KING = "k";
+    private static final String BISHOP = "b";
 
     protected final Color color;
     protected final Position position;
 
-    public AbstractPiece(Color color, Position position) {
+    protected AbstractPiece(Color color, Position position) {
         this.color = color;
         this.position = position;
+    }
+
+    public static Piece of(String symbol, Position position) {
+        String symbolType = symbol.toLowerCase();
+        Color color = getColor(symbol, symbolType);
+        if (PAWN.equals(symbolType)) {
+            return new Pawn(color, position);
+        }
+        if (ROOK.equals(symbolType)) {
+            return new Rook(color, position);
+        }
+        if (KNIGHT.equals(symbolType)) {
+            return new Knight(color, position);
+        }
+        if (QUEEN.equals(symbolType)) {
+            return new Queen(color, position);
+        }
+        if (KING.equals(symbolType)) {
+            return new King(color, position);
+        }
+        if (BISHOP.equals(symbolType)) {
+            return new Bishop(color, position);
+        }
+        throw new IllegalArgumentException("잘못된 입력입니다.");
+    }
+
+    private static Color getColor(String symbol, String symbolType) {
+        if (symbol.equals(symbolType)) {
+            return Color.WHITE;
+        }
+        return Color.BLACK;
     }
 
     @Override
